+++ Package.swift
@@ -8,6 +8,7 @@ let package = Package(
   products: [.executable(name: "HelloMini", targets: ["HelloMini"])],
   targets: [
     .target(name: "MiniCore"),
+    .target(name: "MiniSystem7Theme", dependencies: ["MiniCore", "MiniUI"]),
     .target(name: "MiniScreensaver", dependencies: ["MiniCore", "MiniUI"]),
     .target(name: "MiniToasters", dependencies: ["MiniCore", "MiniUI", "MiniScreensaver"]),
     .target(name: "MiniUI", dependencies: ["MiniCore"], resources: [.copy("Resources")]),
@@ -40,7 +41,7 @@ let package = Package(
         "MiniDesktop", "MiniFinder", "MiniAbout", "MiniActivityMonitor", "MiniClock",
         "MiniControlPanel", "MiniUI", "MiniTeapot", "MiniAquarium", "MiniScrapbook",
         "MiniCalculator", "MiniPuzzle", "MiniChooser", "MiniDiskFirstAid", "MiniWastebasket",
-        "MiniPrintMonitor", "MiniScreensaver", "MiniToasters",
+        "MiniPrintMonitor", "MiniScreensaver", "MiniToasters", "MiniSystem7Theme",
       ]),
     .testTarget(name: "MiniUITests", dependencies: ["MiniUI"], resources: [.process("Resources")]),
     .testTarget(name: "MiniTeapotTests", dependencies: ["MiniTeapot"]),
@@ -55,6 +56,8 @@ let package = Package(
     .testTarget(
       name: "MiniUtilityTests",
       dependencies: ["MiniStorage", "MiniChooser", "MiniBuildCore", "MiniGitHubCI", "MiniGitLabCI"]),
+    .testTarget(
+      name: "MiniSystem7ThemeTests", dependencies: ["MiniSystem7Theme", "MiniUI", "MiniCore"]),
     .testTarget(name: "MiniCoreTests", dependencies: ["MiniCore"]),
     .testTarget(name: "MiniDesktopTests", dependencies: ["MiniDesktop"]),
     .testTarget(name: "MiniFinderTests", dependencies: ["MiniFinder"]),